requirements.txt
torch
transformers
Pillow
opencv-python
fastapi
uvicorn[standard]
opencv-python-headless
numpy
hf_xet